requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib
python-jose
psycopg2
email-validator
bcrypt